Fixed Minimap and Singelton Issues
-Fixed issue with triggers that were triggering multiple times due to the body parts of the player, resulting in multiple small errors throughout the Title
-Fixed Gun reload time from int to float
-Lowered reload time on Gun by default

+++ Assets/Scripts/Weapon/Gun.cs
@@ -10,7 +10,7 @@ public class Gun : MonoBehaviour {
     [Tooltip("Bullets per magazine")]
     public int magSize = 25;
     public int bulletsInMag;
-    public int reloadTime = 5;
+    public float reloadTime = 5;
     private float reloadTimeCounter;
     private Coroutine reloadCoroutine;
     public bool isReloading = false;
